+++ src/pipecatcloud/cli/entry_point.py
@@ -18,6 +18,7 @@ from pipecatcloud.cli.commands.docker import create_docker_command
 from pipecatcloud.cli.commands.organizations import organization_cli
 from pipecatcloud.cli.commands.regions import regions_cli
 from pipecatcloud.cli.commands.secrets import secrets_cli
+from pipecatcloud.cli.commands.spend_limit import spend_limit_cli
 from pipecatcloud.cli.config import config
 from pipecatcloud.exception import ConfigFileError
 
@@ -99,6 +100,7 @@ entrypoint_cli_typer.add_typer(build_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(organization_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(regions_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(secrets_cli, rich_help_panel="Commands")
+entrypoint_cli_typer.add_typer(spend_limit_cli, rich_help_panel="Commands")
 entrypoint_cli_typer.add_typer(agent_cli, rich_help_panel="Commands")
 
 entrypoint_cli = typer.main.get_command(entrypoint_cli_typer)